Load initial story
