+++ packages/wp-story-editor/src/api/utils/index.js
@@ -17,3 +17,4 @@ export { default as flattenFormData } from './flattenFormData';
 export { default as base64Encode } from './base64Encode';
 export { default as snakeCaseToCamelCase } from './snakeCaseToCamelCase';
 export { default as getResourceFromAttachment } from './getResourceFromAttachment';
+export { default as transformGetStoryResponse } from './transformGetStoryResponse';
